+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -1,11 +1,13 @@
 
 -- actual versions:
 -- use "FRAME_VERSION" as mp7_ugt release fw version (used for tag name).
--- mp7_ugt (=FRAME_VERSION): v1.18.0
+-- mp7_ugt (=FRAME_VERSION): v1.19.0
 
--- HB 2021-12-16: v1.18.0 - Cleaned up ../firmware/ngc.
--- v1.17.3: Reorganized ../firmware/ngc, updated add_l1menu_blkmem_files.tcl and gtl_fdl_wrapper_tpl_questa.do
--- v1.17.2: Changed python scripts run_simulation_questa.py, run_simulation_questa_local.py and runIpbbSynth.py
+-- HB 2021-12-22: v1.19.0 - Renamed scripts (no camel case) and updated for different version of Questsim and Vivado
+-- HB 2021-12-18: v1.18.0 - Updated scripts for different version of Questsim and Vivado
+-- HB 2021-12-17: v1.17.3 - Reorganized ../firmware/ngc, updated add_l1menu_blkmem_files.tcl and gtl_fdl_wrapper_tpl_questa.do
+-- HB 2021-12-16: v1.17.2 - Cleaned up ../firmware/ngc.
+
 -- v1.15.4: Changes in frame (v1.2.4) - selector for scouting and no spymem
 -- v1.15.3: Added python script runSynthWorkFlow.py
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
@@ -64,7 +66,7 @@ package gt_mp7_core_pkg is
 -- ==================================================================================================
 -- FRAME version
     constant FRAME_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant FRAME_MINOR_VERSION      : integer range 0 to 255 := 18;
+    constant FRAME_MINOR_VERSION      : integer range 0 to 255 := 19;
     constant FRAME_REV_VERSION        : integer range 0 to 255 := 0;
 	constant FRAME_VERSION : std_logic_vector(31 downto 0) := X"00" &
            std_logic_vector(to_unsigned(FRAME_MAJOR_VERSION, 8)) &