+++ scripts/run_simulation_questa.py
@@ -20,7 +20,8 @@ from threading import Thread
 
 import toolbox as tb
 import xmlmenu
-# from run_compile_simlib import run_compile_simlib
+
+from ignored_algos import IGNORED_ALGOS
 
 # terminal size
 with os.popen('stty size') as fp:
@@ -47,14 +48,24 @@ DefaultQuestaSimLibsPath = os.getenv('UGT_QUESTASIM_LIBS_PATH')
 if not DefaultQuestaSimLibsPath:
     raise RuntimeError('UGT_QUESTASIM_LIBS_PATH is not defined.')
 
+DefaultGitlabUrlIPB = 'https://github.com/ipbus/ipbus-firmware'
+"""Default URL IPB FW repo."""
+
+DefaultIpbbTag = 'v1.4'
+"""Default tag IPB FW repo."""
+
+DefaultGitlabUrlMP7 = 'https://gitlab.cern.ch/hbergaue/mp7'
+"""Default URL MP7 FW repo."""
+
+DefaultMP7Tag = 'mp7fw_v3_0_0_mp7_ugt'
+"""Default tag MP7 FW repo."""
+
 vhdl_snippets_names = [
     'algo_index',
     'gtl_module_instances',
     'gtl_module_signals',
     'ugt_constants'
 ]
-
-url_menu_default = 'https://raw.githubusercontent.com/herbberg/l1menus/master'
 
 DO_FILE = 'gtl_fdl_wrapper.do'
 TB_FILE_TPL = os.path.join('testbench', 'templates', 'gtl_fdl_wrapper_tb_tpl.vhd')
@@ -67,51 +78,17 @@ mp7_tag = 'cactusupgrades'
 
 algonum = 512  # numbers of bits
 
-IGNORED_ALGOS = [
-    'L1_FirstBunchInTrain',
-    'L1_SecondBunchInTrain',
-    'L1_MASSUPT_0_0_10',
-    'L1_MASSUPT_0_0_20',
-    'L1_MASSUPT_0_0_10_open',
-    'L1_MASSUPT_0_0_20_open',
-    'L1_MASSUPT_0_0_10',
-    'L1_MASSUPT_0_0_20',
-    'L1_MASSUPT_5_5_10',
-    'L1_MASSUPT_5_5_20',
-    'L1_MASSUPT_0_0_10_open',
-    'L1_MASSUPT_0_0_20_open',
-    'L1_MASSUPT_5_5_10_open',
-    'L1_MASSUPT_5_5_20_open',
-    'L1_Mu0upt20ip0',
-    'L1_Mu0upt20ip1',
-    'L1_Mu0upt20ip2',
-    'L1_Mu0upt20ip3',
-    'L1_Mu0upt20ip03',
-    'L1_Mu0upt0',
-    'L1_Mu0upt5',
-    'L1_Mu0upt10',
-    'L1_Mu0upt20',
-    'L1_Mu0upt50',
-    'L1_Mu0upt100',
-    'L1_SingleMuOpenupt5',
-    'L1_SingleMuOpenupt20',
-    'L1_SingleMuOpenupt100',
-    'L1_DoubleJet35_Mass_Min450_IsoTau45_RmOvlp',
-    'L1_DoubleJet_80_30_Mass_Min420_IsoTau40_RmOvlp'
-]
-
+# moved IGNORED_ALGOS to "ignored_algos.py" to have an own file for changes in IGNORED_ALGOS list
 
 def run_command(*args):
     command = ' '.join(args)
     logging.info(">$ %s", command)
     os.system(command)
 
-
 def read_file(filename):
     """Returns contents of a file."""
     with open(filename, 'r') as fp:
         return fp.read()
-
 
 def render_template(src, dst, args):
     """Replaces content of file *src* with values of dictionary *args* and writes to file *dst*.
@@ -126,7 +103,6 @@ def render_template(src, dst, args):
     with open(dst, 'w') as dst:
         dst.write(content)
 
-
 def make_testvector(mask, testvectorfile, new_testvector):
     """uses mask of the module, testvector file and the path of the new
     testvector file where the masked testvectors are stored"""
@@ -138,7 +114,6 @@ def make_testvector(mask, testvectorfile, new_testvector):
             colums[-1] = '1' if mask_trigger else '0'
             opf.write(' '.join(colums))
             opf.write('\n')
-
 
 def trigger_list(testvectorfile):
     """makes a list of all triggers in testvectorfile eg. [1,0,0,1,0,1,0,0,1,1,1]"""
@@ -157,7 +132,6 @@ def bitfield(i, n=algonum):
     [0, 1, 0, 1]
     """
     return [int(digit) for digit in '{0:0{1}b}'.format(i, n)][::-1]
-
 
 def run_vsim(vsim, module, msgmode, ini_file):
     """uses class module, arg msgmode and ini file path to start the simulation"""
@@ -199,7 +173,6 @@ def run_vsim(vsim, module, msgmode, ini_file):
 
         logging.info("finished simulating module_{}".format(module._id))
 
-
 def check_algocount(liste):
     """prosseses list so module id is in [0] and trgger count in [1] eg. [1, 255]"""
     aus_liste = []
@@ -220,7 +193,6 @@ def check_multiple(liste):  # checks if multiple triggers in list
 def logging_debug_write(textfile, string):  # output into textfile and if logging.debug true prints on screen
     textfile.write(string + '\n')
     logging.debug(string)
-
 
 class Module(object):
 
@@ -274,27 +246,22 @@ class Module(object):
             '{{gtl_module_instances}}': read_file(os.path.join(src_dir, 'gtl_module_instances.vhd')),
         }
 
-        gtl_fdl_wrapper_dir = os.path.join(uGTalgosPath, 'hdl', 'payload')
-        # gtl_dir = os.path.join(gtl_fdl_wrapper_dir, 'gtl')
-        fdl_dir = os.path.join(gtl_fdl_wrapper_dir, 'fdl')
-        pkg_dir = os.path.join(uGTalgosPath, 'hdl', 'packages')
         # Patch VHDL files
         render_template(
-            os.path.join(fdl_dir, 'algo_mapping_rop_tpl.vhd'),
+            os.path.join(uGTalgosPath, 'hdl', 'payload', 'fdl', 'algo_mapping_rop_tpl.vhd'),
             os.path.join(self.path, 'vhdl', 'algo_mapping_rop.vhd'),
             replace_map
         )
         render_template(
-            os.path.join(pkg_dir, 'fdl_pkg_tpl.vhd'),
+            os.path.join(uGTalgosPath, 'hdl', 'packages', 'fdl_pkg_tpl.vhd'),
             os.path.join(self.path, 'vhdl', 'fdl_pkg.vhd'),
             replace_map
         )
         render_template(
-            os.path.join(gtl_fdl_wrapper_dir, 'gtl_module_tpl.vhd'),
+            os.path.join(uGTalgosPath, 'hdl', 'payload', 'gtl_module_tpl.vhd'),
             os.path.join(self.path, 'vhdl', 'gtl_module.vhd'),
             replace_map
         )
-
 
 def download_file_from_url(url, filename):
     """Download files from URL."""
@@ -311,18 +278,12 @@ def download_file_from_url(url, filename):
     with open(filename, 'w') as fp:
         fp.write(d)
 
+def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questasim, a_questasimlibs, a_output, a_view_wave, a_wlf, a_verbose, a_tv, a_local, a_ignored):
 
-# TODO
-# def run_simulation_questa(a_mp7_tag, a_menu, a_testvector, a_vivado, a_questasim, a_questasimlibs, a_output, a_view_wave, a_wlf, a_verbose):
-def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questasim, a_questasimlibs, a_output, a_view_wave, a_wlf, a_verbose):
-    print("a_mp7_tag: ", a_mp7_tag)
-    print("a_menu: ", a_menu)
-    print("a_url_menu: ", a_url_menu)
-    # print("a_vivado: ", a_vivado)
-    print("a_questasim: ", a_questasim)
-    print("a_questasimlibs: ", a_questasimlibs)
+    print("a_mp7_tag",a_mp7_tag)
+    print("a_ipb_fw_dir",a_ipb_fw_dir)
 
-    sim_dir = os.path.join(os.path.dirname(__file__), '..', 'firmware', 'sim')
+    sim_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'firmware', 'sim')
 
     # Copy modelsim.ini from questasimlib dir to sim dir (to get questasim libs corresponding to Vivado version)
     source_filename = os.path.join(a_questasimlibs, 'modelsim.ini')
@@ -344,29 +305,41 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
     if not os.path.exists(temp_dir):
         os.makedirs(temp_dir)  # makes folders
 
-    logging.info("===========================================================================")
-    logging.info("download XML and testvector file from L1Menu repository ...")
-    # Get l1menus_path for URL
-    # url_menu = "{}/{}".format(url_menu_default, a_menu)
-    url_menu = "{}/{}".format(a_url_menu, a_menu)
-    print("=== url_menu: ", url_menu)
-    xml_name = "{}{}".format(a_menu, '.xml')
-    print("=== xml_name: ", xml_name)
-    menu_filepath = os.path.join(temp_dir, xml_name)
-    print("=== menu_filepath: ", menu_filepath)
-    url = "{}/xml/{}".format(url_menu, xml_name)
-    print("=== url: ", url)
-    download_file_from_url(url, menu_filepath)
-    # Remove "distribution number" from a_menu for testvector file name
-    tv_name = "TestVector_{}{}".format((re.split("-", a_menu)[0]), '.txt')
-    print("=== TV name: ", tv_name)
-    testvector_filepath = os.path.join(temp_dir, tv_name)
-    url = "{}/testvectors/{}".format(url_menu, tv_name)
-    download_file_from_url(url, testvector_filepath)
+    # clone repos of MP7 and IPB-firmware to temp_dir for "non local mode"
+    if not a_local:
+        command = f'bash -c "cd {temp_dir}; git clone {a_mp7_tag}.git mp7; git clone {a_ipb_fw_dir}.git ipbus-firmware"'
+        run_command(command)
 
-    # Get VHDL snippets from menu URL
-    # print "menu_filepath: ", menu_filepath
-    # print "testvector_filepath: ", testvector_filepath
+    if a_local:
+        logging.info("===========================================================================")
+        logging.info("copy XML and testvector file from local L1Menu directory ...")
+        menu_dir = "{}/{}".format(a_url_menu, a_menu)
+        xml_name = "{}{}".format(a_menu, '.xml')
+        menu_filepath = os.path.join(temp_dir, xml_name)
+        menu_xml_file = "{}/xml/{}".format(menu_dir, xml_name)
+        if not a_tv.split(".")[1]:
+            tv_name = "{}{}".format(a_tv, '.txt')
+        else:
+            tv_name = a_tv
+        testvector_file = os.path.join(menu_dir, "testvectors", tv_name)
+        testvector_filepath = os.path.join(temp_dir, tv_name)
+
+        shutil.copyfile(menu_xml_file, menu_filepath)
+        shutil.copyfile(testvector_file, testvector_filepath)
+    else:
+        logging.info("===========================================================================")
+        logging.info("download XML and testvector file from L1Menu repository ...")
+        # Get l1menus_path for URL
+        url_menu = "{}/{}".format(a_url_menu, a_menu)
+        print("url_menu:",url_menu)
+        xml_name = "{}{}".format(a_menu, '.xml')
+        menu_filepath = os.path.join(temp_dir, xml_name)
+        url = "{}/xml/{}".format(url_menu, xml_name)
+        download_file_from_url(url, menu_filepath)
+        tv_name = "TestVector_{}{}".format((re.split("-", a_menu)[0]), '.txt')
+        testvector_filepath = os.path.join(temp_dir, tv_name)
+        url_tv = "{}/testvectors/{}".format(url_menu, tv_name)
+        download_file_from_url(url_tv, testvector_filepath)
 
     timestamp = time.time()  # creates timestamp
     _time = datetime.datetime.fromtimestamp(timestamp).strftime('%Y-%m-%dT%H-%M-%S')  # changes time apperance
@@ -383,13 +356,19 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
         vhdl_src_path = os.path.join('vhdl', f'module_{module._id:d}', 'src')
         temp_dir_module = os.path.join(temp_dir, vhdl_src_path)
         if not os.path.exists(temp_dir_module):
-            os.makedirs(temp_dir_module)
-        for vhdl_name in vhdl_snippets_names:
-            vhdl_name_ext = vhdl_name + ".vhd"
-            vhdl_file_local_path = os.path.join(temp_dir_module, vhdl_name_ext)
-            vhdl_file_path = os.path.join(vhdl_src_path, vhdl_name_ext)
-            url = "{}/{}".format(url_menu, vhdl_file_path)
-            download_file_from_url(url, vhdl_file_local_path)
+            os.makedirs(temp_dir_module)  # makes folders
+            for vhdl_name in vhdl_snippets_names:
+                vhdl_name_ext = vhdl_name + ".vhd"
+                vhdl_file_local_path = os.path.join(temp_dir_module, vhdl_name_ext)
+                vhdl_file_path = os.path.join(vhdl_src_path, vhdl_name_ext)
+                if a_local:
+                    vhdl_file_source = "{}/{}".format(menu_dir, vhdl_file_path)
+                    command = 'bash -c "cd; cp {vhdl_file_source} {vhdl_file_local_path}"'.format(**locals())
+                    run_command(command)
+                else:
+                    url = "{}/{}".format(url_menu, vhdl_file_path)
+                    print("url:",url)
+                    download_file_from_url(url, vhdl_file_local_path)
 
     if not os.path.exists(menu_filepath):
         raise RuntimeError('Missing %s File' % menu_filepath)
@@ -402,7 +381,7 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
 
     ini_file = os.path.join(sim_dir, INI_FILE)
 
-    logging.info('Creating Modules and Masks...')
+    logging.info('creating Modules and Masks...')
 
     for module in modules:  # gives each module the information
         module_id = 'module_%d' % module._id
@@ -417,12 +396,18 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
 
         logging.debug('Module_%d created at %s' % (module._id, base_dir))
 
-        # module.make_files(sim_dir, a_view_wave, a_mp7_tag, a_menu)  # sim_dir, view_wave, mp7_tag, menu_path
-        module.make_files(sim_dir, a_view_wave, a_mp7_tag, temp_dir, a_ipb_fw_dir)  # sim_dir, view_wave, mp7_tag, temp_dir
+        if a_local:
+            mp7 = a_mp7_tag
+            ipb_fw = a_ipb_fw_dir
+        else:
+            mp7 = os.path.join(temp_dir, 'mp7')
+            ipb_fw = os.path.join(temp_dir, 'ipbus-firmware')
+
+        module.make_files(sim_dir, a_view_wave, mp7, temp_dir, ipb_fw)  # sim_dir, view_wave, mp7_tag, temp_dir
 
     questasim_path = os.path.join(QuestaSimPath, 'questasim')
 
-    logging.info('finished creating modules and masks')
+    logging.info('finished creating Modules and Masks')
     logging.info("===========================================================================")
     logging.info('starting simulations with Questa Simulator version %s (from directory %s)', a_questasim, questasim_path)
 
@@ -475,6 +460,8 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
     handler.setLevel(logging.DEBUG)
     sum_log.addHandler(handler)
 
+    if a_local:
+        sum_log.info('Test vector file name: {}'.format(tv_name))
     sum_log.info("|-----|-----|------------------------------------------------------------------|--------|--------|--------|")
     sum_log.info("| Mod | Idx | Name of algorithm                                                | l1a.tv | l1a.hw | Result |")
     sum_log.info("|-----|-----|------------------------------------------------------------------|--------|--------|--------|")
@@ -484,7 +471,7 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
     success = True
     for algo in algorithms:
         result = ok_green
-        if algo.name in IGNORED_ALGOS:
+        if algo.name in IGNORED_ALGOS and a_ignored:
             result = ignore_yellow
         # checks if algorithm trigger count is equal in both hardware and testvectors
         elif algos_tv[algo.index][0][1] != algos_sim[algo.index][0][1]:
@@ -556,21 +543,27 @@ def run_simulation_questa(a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir, a_questas
         logging.info("===========================================================================")
         exit(1)
 
-
 def parse_args():
     parser = argparse.ArgumentParser()
-    parser.add_argument('menu', type=tb.menuname_t, help="menu name (eg. 'L1Menu_Collisions2018_v2_1_0-d1')")
-    parser.add_argument('--url', default=url_menu_default, help="URL of menu")
-    parser.add_argument('--mp7_tag', required=True, type=os.path.abspath, help="local path to MP7 tag (checkout tag before running simulation)")
-    parser.add_argument('--ipb_fw_dir', required=True, type=os.path.abspath, help="local path to IPBus firmware directory")
+    parser.add_argument('menu', type=tb.menuname_t, help="menu name (eg. 'L1Menu_Collisions2020_v0_8_1-d1')")
+    parser.add_argument('--menu_url', help="URL of L1Menu")
+    parser.add_argument('--mp7_url', default=DefaultGitlabUrlMP7, help="MP7 repo (default is '{}')".format(DefaultGitlabUrlMP7))
+    parser.add_argument('--mp7_repo_tag', default=DefaultMP7Tag, help="MP7 repo tag (default is '{}')".format(DefaultMP7Tag))
+    parser.add_argument('--ipb_fw_url', default=DefaultGitlabUrlIPB, help="IPBus firmware repo (default is '{}')".format(DefaultGitlabUrlIPB))
+    parser.add_argument('--ipb_fw_repo_tag', default=DefaultIpbbTag, help="IPBus firmware repo tag (default is '{}')".format(DefaultIpbbTag))
+    parser.add_argument('--menu_local', type=os.path.abspath, help="local path to L1Menu")
+    parser.add_argument('--mp7_local', type=os.path.abspath, help="local path to MP7 tag (checkout tag before running simulation)")
+    parser.add_argument('--ipb_fw_local', type=os.path.abspath, help="local path to IPBus firmware directory (checkout tag before running simulation)")
+    parser.add_argument('--local', action='store_true', default=False, help="running simulation with Questa simulator in local mode")
+    parser.add_argument('--tv', help="Test vector name (needed only with 'local')")
+    parser.add_argument('--ignored', action='store_true', default=False, help="using IGNORED_ALGOS for error checks")
     parser.add_argument('--questasim', type=tb.questasim_t, default=DefaultQuestasimVersion, help="Questasim version (default is {})".format(DefaultQuestasimVersion))
     parser.add_argument('--questasimlibs', default=DefaultQuestaSimLibsPath, help="Questasim Vivado libraries directory name (default is {})".format(DefaultQuestaSimLibsPath))
-    parser.add_argument('--output', metavar='path', help='', type=os.path.abspath)
+    parser.add_argument('--output', metavar='path', type=os.path.abspath, help='path to output directory')
     parser.add_argument('--view-wave', action='store_true', help="shows the waveform")
     parser.add_argument('--wlf', action='store_true', help="no console transcript info, warning and error messages (transcript output to vsim.wlf)")
     parser.add_argument('-v', '--verbose', action='store_const', const=logging.DEBUG, help="enables debug prints to console", default=logging.INFO)
     return parser.parse_args()
-
 
 def main():
     args = parse_args()
@@ -578,10 +571,28 @@ def main():
     # Setup console logging
     logging.basicConfig(format='%(levelname)s: %(message)s', level=logging.INFO)
 
-    # TODO
-    # run_simulation_questa(args.mp7_tag, args.menu, args.url, args.ipb_fw_dir, args.vivado, args.questasim, args.questasimlibs, args.output, args.view_wave, args.wlf, args.verbose)
-    run_simulation_questa(args.mp7_tag, args.menu, args.url, args.ipb_fw_dir, args.questasim, args.questasimlibs, args.output, args.view_wave, args.wlf, args.verbose)
+    if args.local:
+        menu_p = args.menu_local
+        if not menu_p:
+            raise RuntimeError('argument --menu_local is not set.')
+        mp7 = args.mp7_local
+        if not mp7:
+            raise RuntimeError('argument --mp7_local is not set.')
+        ipb_fw = args.ipb_fw_local
+        if not ipb_fw:
+            raise RuntimeError('argument --ipb_fw_local is not set.')
+        tv = args.tv
+        if not tv:
+            raise RuntimeError('argument --tv is not set.')
+    else:
+        menu_p = args.menu_url
+        if not menu_p:
+            raise RuntimeError('argument --menu_url is not set.')
+        mp7 = args.mp7_url
+        ipb_fw = args.ipb_fw_url
+        tv =''
 
+    run_simulation_questa(mp7, args.menu, menu_p, ipb_fw, args.questasim, args.questasimlibs, args.output, args.view_wave, args.wlf, args.verbose, tv, args.local, args.ignored)
 
 if __name__ == '__main__':
     main()