+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.19.0] - 2022-01-10
+### Comment
+
+- mp7_ugt firmware release v1.19.0
+- renamed scripts (no camel case) and updated for different version of Questsim and Vivado
+
+### Changed
+- script files:
+  - scripts/run_simulation_questa.py
+  - scripts/build_report.py
+  - scripts/check_synth_ipbb.py
+  - scripts/fwpacker_ipbb.py
+  - scripts/run_synth_ipbb.py
+  - scripts/run_synth_workflow.py
+- source files:
+  - ../packages/gt_mp7_core_pkg.vhd
+- readme.md
+
 ## [v1.18.0] - 2021-12-18
 ### Comment